+++ src/toisto/model/progress.py
@@ -25,13 +25,15 @@ class Progress:
     def next_quiz(self, quizzes: list[Quiz]) -> Quiz | None:
         """Return the next quiz."""
         if eligible_quizzes := [quiz for quiz in quizzes if self.is_eligible(quiz)]:
+            if len(eligible_quizzes) > 1 and self.current_quiz in eligible_quizzes:
+                eligible_quizzes.remove(self.current_quiz)
             self.current_quiz = random.choice(eligible_quizzes)
             return self.current_quiz
         return None
-
+#
     def is_eligible(self, quiz: Quiz) -> bool:
         """Is the quiz eligible."""
-        return quiz != self.current_quiz and not self.get_progress(quiz).is_silenced()
+        return not self.get_progress(quiz).is_silenced()
 
     def as_dict(self) -> dict[str, dict[str, int | str]]:
         """Return the progress as dict."""